更新包版本和数据类型调整
更新了多个项目文件中的包版本：
- `DH.AspNetCore.csproj` 中的 `Pek.AspNetCore` 包版本从 `4.0.2024.1126-beta0909` 更新到 `4.0.2024.1126-beta1212`。
- `DH.SearchEngine.csproj` 和 `DH.Utils.csproj` 中的 `DH.NCode` 包版本从 `4.0.2024.1126-beta0203` 更新到 `4.0.2024.1127-beta0621`。
- `version.props` 中的 `VersionPrefixBeta` 从 `00012` 更新到 `00013`。

将“网站开启关闭状态”的数据类型从 `Byte` 修改为 `Int16`：
- `DG.htm`
- `ISiteInfo.cs`
- `Model.xml`
- `SiteInfoModel.cs`
- `站点基础信息.cs`（包括 `_Status` 字段和 `case "Status"` 的处理逻辑）

+++ src/DH.Core/Entity/Models/SiteInfoModel.cs
@@ -90,7 +90,7 @@ public partial class SiteInfoModel
     public String SiteCopyright { get; set; }
 
     /// <summary>网站开启关闭状态</summary>
-    public Byte Status { get; set; }
+    public Int16 Status { get; set; }
 
     /// <summary>如果状态关闭，请输入关闭网站原因</summary>
     public String CloseInfo { get; set; }
